[Refactor] Split connector data source provider
