+++ be/src/connector/CMakeLists.txt
@@ -17,6 +17,7 @@ set(LIBRARY_OUTPUT_PATH "${BUILD_DIR}/src/connector")
 ADD_BE_LIB(Connector
         connector.cpp
         data_source.cpp
+        data_source_provider.cpp
         connector_registry.cpp
         connector_registry_init.cpp
         hive_connector.cpp
