+++ be/src/connector/connector.h
@@ -18,7 +18,7 @@
 #include <string>
 
 #include "connector/connector_chunk_sink.h"
-#include "connector/data_source.h"
+#include "connector/data_source_provider.h"
 #include "gen_cpp/PlanNodes_types.h"
 
 namespace starrocks {
