+++ be/src/connector/data_source.cpp
@@ -14,11 +14,6 @@
 
 #include "connector/data_source.h"
 
-#include <algorithm>
-#include <functional>
-
-#include "exec/pipeline/scan/morsel.h"
-#include "exec/pipeline/scan/scan_morsel.h"
 #include "exec/runtime_filter/runtime_filter_helper.h"
 #include "exprs/expr_context.h"
 #include "runtime/runtime_state.h"
@@ -63,44 +58,6 @@ Status DataSource::parse_runtime_filters(RuntimeState* state) {
 void DataSource::update_profile(const Profile& profile) {
     RuntimeProfile::Counter* mem_alloc_failed_counter = ADD_COUNTER(_runtime_profile, "MemAllocFailed", TUnit::UNIT);
     COUNTER_UPDATE(mem_alloc_failed_counter, profile.mem_alloc_failed_count);
-}
-
-StatusOr<pipeline::MorselQueuePtr> DataSourceProvider::convert_scan_range_to_morsel_queue(
-        const std::vector<TScanRangeParams>& scan_ranges, int node_id, int32_t pipeline_dop,
-        bool enable_tablet_internal_parallel, TTabletInternalParallelMode::type tablet_internal_parallel_mode,
-        size_t num_total_scan_ranges, size_t scan_parallelism) {
-    peek_scan_ranges(scan_ranges);
-
-    pipeline::Morsels morsels;
-    bool has_more_morsel = false;
-    pipeline::ScanMorsel::build_scan_morsels(node_id, scan_ranges, accept_empty_scan_ranges(), &morsels,
-                                             &has_more_morsel);
-
-    if (partition_order_hint().has_value()) {
-        bool asc = partition_order_hint().value();
-        std::stable_sort(morsels.begin(), morsels.end(), [asc](auto& l, auto& r) {
-            auto l_partition_id = down_cast<pipeline::ScanMorsel*>(l.get())->partition_id();
-            auto r_partition_id = down_cast<pipeline::ScanMorsel*>(r.get())->partition_id();
-            if (asc) {
-                return std::less()(l_partition_id, r_partition_id);
-            } else {
-                return std::greater()(l_partition_id, r_partition_id);
-            }
-        });
-    }
-
-    if (output_chunk_by_bucket()) {
-        std::stable_sort(morsels.begin(), morsels.end(), [](auto& l, auto& r) {
-            return down_cast<pipeline::ScanMorsel*>(l.get())->owner_id() <
-                   down_cast<pipeline::ScanMorsel*>(r.get())->owner_id();
-        });
-    }
-
-    auto morsel_queue = std::make_unique<pipeline::DynamicMorselQueue>(std::move(morsels), has_more_morsel);
-    if (scan_parallelism > 0) {
-        morsel_queue->set_max_degree_of_parallelism(scan_parallelism);
-    }
-    return morsel_queue;
 }
 
 } // namespace starrocks::connector
